EditUserForm: extract `withUpdateUser` hoc
EditUserForm: improve general state handling
